+++ Changelog.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+## 2.18.3
+### Fixed
+ - Changed Object.assign to lodash assign for compatibility.
+
+## 2.18.2 
+### Added
+ -  Allow interpolation of file directories for uploads.
+
+### Changed
+ - Select resource components use Formio provider instead of $http.get to allow offline compatibility.
+ 
+### Fixed
+ - Fixed false value for checkbox.
+
 ## 2.18.1
 ### Added
  - Add setProjectUrl and getProjectUrl options to FormioProvider.